Implemented feedback on events functionality
