+++ src/Observer/ManageLayoutBlocks.php
@@ -10,10 +10,8 @@ use Magento\Framework\App\Request\Http;
 use Magento\Framework\Event\ObserverInterface;
 use Magento\Framework\Event\Observer;
 use Magento\Framework\Exception\LocalizedException;
-use Magento\Framework\View\DesignInterface;
 use Magento\Framework\View\Layout;
 use Tweakwise\TweakwiseJs\Model\Config;
-use Tweakwise\TweakwiseJs\Model\Enum\SearchType;
 
 class ManageLayoutBlocks implements ObserverInterface
 {
@@ -23,21 +21,14 @@ class ManageLayoutBlocks implements ObserverInterface
     private Layout $layout;
 
     /**
-     * @var bool|null
-     */
-    private ?bool $isHyva = null;
-
-    /**
      * @param Http $request
      * @param Config $config
      * @param Resolver $layerResolver
-     * @param DesignInterface $viewDesign
      */
     public function __construct(
         private readonly Http $request,
         private readonly Config $config,
         private readonly Resolver $layerResolver,
-        private readonly DesignInterface $viewDesign
     ) {
     }
 
@@ -54,34 +45,11 @@ class ManageLayoutBlocks implements ObserverInterface
 
         $this->layout = $observer->getLayout();
 
-        $this->addDefaultHandle();
-        $this->addSearchHandle();
-
-        if ($this->shouldAddAddToCartWishlistHandle()) {
-            $this->addAddToCartWishlistHandle();
-        }
-
         if (!$this->isCategoryPage() || !$this->shouldShowTweakwiseJsCategoryViewBlock()) {
             return;
         }
 
         $this->addMerchandisingHandle();
-    }
-
-    /**
-     * @return void
-     */
-    private function addDefaultHandle(): void
-    {
-        $this->layout->getUpdate()->addHandle('tweakwisejs_default');
-    }
-
-    /**
-     * @return void
-     */
-    private function addSearchHandle(): void
-    {
-        $this->layout->getUpdate()->addHandle('tweakwisejs_search');
     }
 
     /**
@@ -93,27 +61,11 @@ class ManageLayoutBlocks implements ObserverInterface
     }
 
     /**
-     * @return void
-     */
-    private function addAddToCartWishlistHandle(): void
-    {
-        $this->layout->getUpdate()->addHandle($this->getHandleName('tweakwisejs_add_to'));
-    }
-
-    /**
      * @return bool
      */
     private function isCategoryPage(): bool
     {
         return $this->request->getFullActionName() === 'catalog_category_view';
-    }
-
-    /**
-     * @return bool
-     */
-    private function isSearchResultsPage(): bool
-    {
-        return $this->request->getFullActionName() === 'catalogsearch_results_index';
     }
 
     /**
@@ -136,46 +88,5 @@ class ManageLayoutBlocks implements ObserverInterface
         }
 
         return true;
-    }
-
-    /**
-     * @return bool
-     */
-    private function shouldAddAddToCartWishlistHandle(): bool
-    {
-        return $this->isCategoryPage() ||
-            $this->isSearchResultsPage() ||
-            $this->config->getSearchType()->value === SearchType::INSTANT_SEARCH->value;
-    }
-
-    /**
-     * @return bool
-     */
-    private function isHyva(): bool
-    {
-        if ($this->isHyva !== null) {
-            return $this->isHyva;
-        }
-
-        $theme = $this->viewDesign->getDesignTheme();
-        while ($theme) {
-            if (str_starts_with($theme->getCode(), 'Hyva/')) {
-                $this->isHyva = true;
-                return $this->isHyva;
-            }
-            $theme = $theme->getParentTheme();
-        }
-
-        $this->isHyva = false;
-        return $this->isHyva;
-    }
-
-    /**
-     * @param string $handle
-     * @return string
-     */
-    private function getHandleName(string $handle): string
-    {
-        return $this->isHyva() ? sprintf('hyva_%s', $handle) : $handle;
     }
 }
