+++ src/ViewModel/Event.php
@@ -7,45 +7,24 @@ namespace Tweakwise\TweakwiseJs\ViewModel;
 use Magento\Checkout\Model\Session;
 use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Framework\Serialize\Serializer\Json;
-use Magento\Framework\View\Element\Block\ArgumentInterface;
 use Tweakwise\TweakwiseJs\Helper\Data;
 use Tweakwise\TweakwiseJs\Model\Config;
 
-class Event implements ArgumentInterface
+class Event extends Base
 {
     /**
-     * @param Data $dataHelper
      * @param Config $config
+     * @param Data $dataHelper
      * @param Session $checkoutSession
      * @param Json $jsonSerializer
      */
     public function __construct(
-        private readonly Data $dataHelper,
-        private readonly Config $config,
+        Config $config,
+        Data $dataHelper,
         private readonly Session $checkoutSession,
         private readonly Json $jsonSerializer
     ) {
-    }
-
-    /**
-     * @param int $productId
-     * @return string
-     */
-    public function getTweakwiseProductId(int $productId): string
-    {
-        try {
-            return $this->dataHelper->getTweakwiseId($productId);
-        } catch (NoSuchEntityException $e) {
-            return '0';
-        }
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getInstanceKey(): ?string
-    {
-        return $this->config->getInstanceKey();
+        parent::__construct($config, $dataHelper);
     }
 
     /**
