+++ src/view/frontend/layout/catalogsearch_results_index.xml
@@ -9,19 +9,6 @@
                     <argument name="text" xsi:type="string"><![CDATA[&nbsp;]]></argument>
                 </arguments>
             </block>
-            <block class="Magento\Framework\View\Element\Template"
-                   name="tweakwise-js-event-default"
-                   template="Tweakwise_TweakwiseJs::js/event/default.phtml"
-                   ifconfig="tweakwise/tweakwisejs/events/enabled">
-                <arguments>
-                    <argument name="view_model" xsi:type="object">Tweakwise\TweakwiseJs\ViewModel\Event</argument>
-                </arguments>
-            </block>
-            <block class="Magento\Framework\View\Element\Template"
-                   name="tweakwise-js-event-search"
-                   template="Tweakwise_TweakwiseJs::js/event/search-event.phtml"
-                   ifconfig="tweakwise/tweakwisejs/events/enabled">
-            </block>
         </referenceContainer>
         <move element="customer.customer.data" destination="columns"/>
         <move element="customer.section.config" destination="columns"/>
