--- src/view/frontend/layout/tweakwisejs_add_to.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0"?>
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <body>
-        <referenceContainer name="page.wrapper">
-            <block class="Magento\Framework\View\Element\Template" name="tweakwise-js-add-to-js"
-                   template="Tweakwise_TweakwiseJs::js/category/add-to.phtml">
-                <arguments>
-                    <argument name="view_model" xsi:type="object">Tweakwise\TweakwiseJs\ViewModel\Merchandising
-                    </argument>
-                </arguments>
-            </block>
-        </referenceContainer>
-    </body>
-</page>
