--- src/view/frontend/layout/tweakwisejs_default.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0"?>
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <body>
-        <referenceBlock name="after.body.start">
-            <block class="Magento\Framework\View\Element\Template" name="tweakwise-js-default"
-                   template="Tweakwise_TweakwiseJs::js/default.phtml"/>
-        </referenceBlock>
-    </body>
-</page>
