--- src/view/frontend/layout/tweakwisejs_search.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0"?>
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <body>
-        <referenceBlock name="after.body.start">
-            <block class="Magento\Framework\View\Element\Template" name="tweakwise-js-search"
-                   template="Tweakwise_TweakwiseJs::js/search.phtml">
-                <arguments>
-                    <argument name="view_model" xsi:type="object">Tweakwise\TweakwiseJs\ViewModel\Search</argument>
-                </arguments>
-            </block>
-        </referenceBlock>
-    </body>
-</page>
